+++ src/Kupo/App/ChainSync/Ogmios.hs
@@ -12,9 +12,6 @@ module Kupo.App.ChainSync.Ogmios
 
 import Kupo.Prelude
 
-import Control.Exception.Safe
-    ( MonadThrow
-    )
 import Kupo.App.Mailbox
     ( Mailbox
     , putHighFrequencyMessage
@@ -22,6 +19,9 @@ import Kupo.App.Mailbox
     )
 import Kupo.Control.MonadSTM
     ( MonadSTM (..)
+    )
+import Kupo.Control.MonadThrow
+    ( MonadThrow (..)
     )
 import Kupo.Data.Cardano
     ( Point